# services/backend/public/api-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Svetofor Furniture API",
    "version": "1.0.0",
    "description": "API для интернет-магазина мебели"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Локальный сервер"
    },
    {
      "url": "https://demo1.site.zone",
      "description": "Демо-стенд (бэкенд)"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "svetofor-session",
        "description": "Основная сессионная кука Laravel. Ставится сервером при первом запросе. Фронт обязан слать все запросы с credentials:\"include\" (withCredentials). Аутентификация — session/cookie-based (Sanctum SPA), НЕ Bearer-токен."
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "CSRF-токен. Читается из куки XSRF-TOKEN и передаётся заголовком для небезопасных методов (POST/PUT/DELETE) на sanctum-эндпоинтах."
      }
    },
    "schemas": {
...
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "product_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "total": {
            "type": "number"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "is_variant": {
            "type": "boolean",
            "nullable": true
          },
          "variation_attributes": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "delivery_days": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/cart": {
      "get": {
        "operationId": "getCart",
        "summary": "Получить содержимое корзины",
        "description": "Возвращает все товары в корзине с ценами и количеством",
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "region_id",
            "in": "query",
            "description": "ID региона доставки для расчета цен",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "example": 558
          }
        ],
        "responses": {
          "200": {
            "description": "Содержимое корзины",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CartItem"
                      }
                    },
                    "subtotal": {
                      "type": "number"
                    },
                    "item_count": {
                      "type": "integer"
                    },
                    "is_empty": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addToCart",
        "summary": "Добавить товар в корзину",
        "description": "Добавляет товар в корзину. Поддерживает вариации товаров (цвет, размер). Автоматически корректирует количество с учетом остатков. Если товар уже в корзине, количество прибавляется к той же позиции (id сохраняется); ограничение — в limit_reason.",
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product_id": {
                    "type": "integer",
                    "description": "exists:products"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  },
                  "variation_attributes": {
                    "type": "array",
                    "description": "Атрибуты вариации (для вариативных товаров)",
                    "items": {
                      "type": "object",
                      "properties": {
                        "attribute_slug": {
                          "type": "string"
                        },
                        "value_slug": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "attribute_slug",
                        "value_slug"
                      ]
                    }
                  }
                },
                "required": [
                  "product_id"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар добавлен в корзину",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Товар добавлен в корзину"
                    },
                    "item": {
                      "type": "object"
                    },
                    "requested_quantity": {
                      "type": "integer",
                      "example": 1
                    },
                    "added_quantity": {
                      "type": "integer",
                      "example": 1
                    },
                    "was_adjusted": {
                      "type": "boolean",
                      "example": false
                    },
                    "limit_reason": {
                      "type": "string",
                      "nullable": true,
                      "enum": [
                        "stock",
                        "max_per_order"
                      ],
                      "description": "Почему количество урезано: stock — по остатку, max_per_order — лимит 100 шт. на позицию; null — сохранено запрошенное количество",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации или недостаточно товара на складе"
          },
          "404": {
            "description": "Вариация товара не найдена"
          }
        }
      },
      "delete": {
        "operationId": "clearCart",
        "summary": "Очистить корзину",
        "description": "Удаляет все товары из корзины",
        "tags": [
          "Cart"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Корзина очищена",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Корзина очищена"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}